[Fix] "info" in append callback
Shame on me! The weights stored in the lib memory is not the same as the
returned from servers. For now, just let the "append" function continue
normally and make a new request. It will rely on browser cache to
prevent unuseful requests. Actually, if it will always request, there is
no point storing previous requests except for the `link` tag.